modularized the file structure of veilid
