+++ packages/grid/veilid/server/veilid_core.py
@@ -1,11 +1,4 @@
-# stdlib
-import base64
-from collections.abc import Callable
-import json
-import lzma
-
 # third party
-import httpx
 from loguru import logger
 import veilid
 from veilid import KeyPair
@@ -13,100 +6,17 @@ from veilid import Sequencing
 from veilid import Stability
 from veilid import TypedKey
 from veilid import ValueData
-from veilid import VeilidUpdate
 from veilid.json_api import _JsonRoutingContext
 from veilid.json_api import _JsonVeilidAPI
 from veilid.types import RouteId
 
 # relative
-from .constants import HOST
-from .constants import PORT
 from .constants import USE_DIRECT_CONNECTION
+from .veilid_connection import get_routing_context
+from .veilid_connection import get_veilid_conn
 from .veilid_db import load_dht_key
 from .veilid_db import store_dht_key
 from .veilid_db import store_dht_key_creds
-
-
-async def main_callback(update: VeilidUpdate) -> None:
-    # TODO: Handle other types of network events like
-    # when our private route goes
-    if update.kind == veilid.VeilidUpdateKind.APP_MESSAGE:
-        logger.info(f"Received App Message: {update.detail.message}")
-
-    elif update.kind == veilid.VeilidUpdateKind.APP_CALL:
-        logger.info(f"Received App Call: {update.detail.message}")
-        message: dict = json.loads(update.detail.message)
-
-        async with httpx.AsyncClient() as client:
-            data = message.get("data", None)
-            # TODO: can we optimize this?
-            # We encode the data to base64,as while sending
-            # json expects valid utf-8 strings
-            if data:
-                message["data"] = base64.b64decode(data)
-            response = await client.request(
-                method=message.get("method"),
-                url=message.get("url"),
-                data=message.get("data", None),
-                params=message.get("params", None),
-                json=message.get("json", None),
-            )
-
-        async with await get_veilid_conn() as conn:
-            compressed_response = lzma.compress(response.content)
-            logger.info(f"Compression response size: {len(compressed_response)}")
-            await conn.app_call_reply(update.detail.call_id, compressed_response)
-
-
-async def noop_callback(update: VeilidUpdate) -> None:
-    pass
-
-
-async def get_veilid_conn(
-    host: str = HOST, port: int = PORT, update_callback: Callable = noop_callback
-) -> _JsonVeilidAPI:
-    return await veilid.json_api_connect(
-        host=host, port=port, update_callback=update_callback
-    )
-
-
-async def get_routing_context(conn: _JsonVeilidAPI) -> _JsonRoutingContext:
-    if USE_DIRECT_CONNECTION:
-        return await (await conn.new_routing_context()).with_safety(
-            veilid.SafetySelection.unsafe(veilid.Sequencing.ENSURE_ORDERED)
-        )
-    else:
-        return await (await conn.new_routing_context()).with_sequencing(
-            veilid.Sequencing.ENSURE_ORDERED
-        )
-
-
-class VeilidConnectionSingleton:
-    _instance = None
-
-    def __new__(cls) -> "VeilidConnectionSingleton":
-        if cls._instance is None:
-            cls._instance = super().__new__(cls)
-            cls._instance._connection = None
-        return cls._instance
-
-    def __init__(self) -> None:
-        self._connection: _JsonVeilidAPI | None = None
-
-    @property
-    def connection(self) -> _JsonVeilidAPI | None:
-        return self._connection
-
-    async def initialize_connection(self) -> None:
-        if self._connection is None:
-            self._connection = await get_veilid_conn(update_callback=main_callback)
-            logger.info("Connected to Veilid")
-
-    async def release_connection(self) -> None:
-        if self._connection is not None:
-            await self._connection.release()
-            logger.info("Disconnected  from Veilid")
-            self._connection = None
 
 
 async def create_private_route(
